refactor(contracts): close the implicit re-export of the slug parser (#725)
The move left parse_github_repo_slug importable from its old Cortex home
under its public name, so the cross-layer path kept resolving. Cortex now
imports it under a private alias and the old path raises ImportError.

Move the parser's tests to tests/test_contracts_github.py, following the
function to its owner, and add a regression test so an unaliased import
fails the suite instead of silently reopening the path.
